+++ packages/world-vercel/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @workflow/world-vercel
 
+## 4.0.1-beta.4
+
+### Patch Changes
+
+- 20d51f0: Enforce the Vercel Queue max visibility limit
+- Updated dependencies [796fafd]
+- Updated dependencies [20d51f0]
+- Updated dependencies [70be894]
+  - @workflow/errors@4.0.1-beta.2
+  - @workflow/world@4.0.1-beta.3
+
+## 4.0.1-beta.3
+
+### Patch Changes
+
+- e367046: Allow setting baseUrl and token for queue service
+
 ## 4.0.1-beta.2
 
 ### Patch Changes